refactor: moved functionality unique to MakeDecision to its own gateway.
