+++ data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/config/MakeDecisionConfig.java
@@ -5,16 +5,16 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import tools.jackson.databind.ObjectMapper;
 import uk.gov.justice.laa.dstew.access.controller.application.MakeDecisionCommandMapper;
+import uk.gov.justice.laa.dstew.access.infrastructure.jpa.makedecision.MakeDecisionApplicationJpaGateway;
 import uk.gov.justice.laa.dstew.access.infrastructure.jpa.makedecision.MakeDecisionCertificateJpaGateway;
 import uk.gov.justice.laa.dstew.access.infrastructure.jpa.makedecision.MakeDecisionGatewayMapper;
 import uk.gov.justice.laa.dstew.access.infrastructure.jpa.makedecision.MakeDecisionProceedingJpaGateway;
-import uk.gov.justice.laa.dstew.access.infrastructure.jpa.shared.application.ApplicationGatewayMapper;
-import uk.gov.justice.laa.dstew.access.infrastructure.jpa.shared.application.ApplicationJpaGateway;
 import uk.gov.justice.laa.dstew.access.repository.ApplicationRepository;
 import uk.gov.justice.laa.dstew.access.repository.CertificateRepository;
 import uk.gov.justice.laa.dstew.access.repository.ProceedingRepository;
 import uk.gov.justice.laa.dstew.access.service.domainevents.SaveDomainEventService;
 import uk.gov.justice.laa.dstew.access.usecase.makedecision.MakeDecisionUseCase;
+import uk.gov.justice.laa.dstew.access.usecase.shared.infrastructure.ApplicationGateway;
 
 /** Spring configuration that wires all beans for the makeDecision use case. */
 @Configuration
@@ -27,8 +27,8 @@ public class MakeDecisionConfig {
   private final SaveDomainEventService saveDomainEventService;
   private final ObjectMapper objectMapper;
 
-  /** Injected from {@link CreateApplicationConfig} — the single ApplicationGatewayMapper bean. */
-  private final ApplicationGatewayMapper applicationGatewayMapper;
+  /** Injected shared gateway — used for read operations (findByApplicationId). */
+  private final ApplicationGateway applicationGateway;
 
   /** Creates the certificate gateway mapper bean. */
   @Bean
@@ -55,23 +55,29 @@ public class MakeDecisionConfig {
     return new MakeDecisionCommandMapper(objectMapper);
   }
 
+  /** Creates the application JPA gateway bean for decision-update operations. */
+  @Bean
+  public MakeDecisionApplicationJpaGateway makeDecisionApplicationGateway(
+      MakeDecisionGatewayMapper makeDecisionGatewayMapper) {
+    return new MakeDecisionApplicationJpaGateway(applicationRepository, makeDecisionGatewayMapper);
+  }
+
   /**
-   * Creates the {@link MakeDecisionUseCase} bean. The {@link ApplicationJpaGateway} for this use
-   * case is constructed locally to avoid exposing a second {@code ApplicationGateway} bean in the
-   * context (which would cause ambiguity for other configs that inject by type).
+   * Creates the {@link MakeDecisionUseCase} bean.
    *
+   * @param makeDecisionApplicationGateway the gateway for decision-update operations
    * @param makeDecisionCertificateGateway the certificate gateway
    * @param makeDecisionProceedingGateway the proceeding gateway
    * @return a fully configured use case
    */
   @Bean
   public MakeDecisionUseCase makeDecisionUseCase(
+      MakeDecisionApplicationJpaGateway makeDecisionApplicationGateway,
       MakeDecisionCertificateJpaGateway makeDecisionCertificateGateway,
       MakeDecisionProceedingJpaGateway makeDecisionProceedingGateway) {
-    ApplicationJpaGateway applicationGateway =
-        new ApplicationJpaGateway(applicationRepository, applicationGatewayMapper);
     return new MakeDecisionUseCase(
         applicationGateway,
+        makeDecisionApplicationGateway,
         makeDecisionCertificateGateway,
         makeDecisionProceedingGateway,
         saveDomainEventService);
