+++ data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/infrastructure/jpa/shared/application/ApplicationGatewayMapper.java
@@ -1,12 +1,8 @@
 package uk.gov.justice.laa.dstew.access.infrastructure.jpa.shared.application;
 
-import java.time.Instant;
 import java.util.LinkedHashSet;
-import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
-import java.util.UUID;
 import java.util.stream.Collectors;
 import uk.gov.justice.laa.dstew.access.domain.ApplicationDomain;
 import uk.gov.justice.laa.dstew.access.domain.DecisionDomain;
@@ -20,10 +16,8 @@ import uk.gov.justice.laa.dstew.access.entity.MeritsDecisionEntity;
 import uk.gov.justice.laa.dstew.access.entity.ProceedingEntity;
 import uk.gov.justice.laa.dstew.access.model.ApplicationStatus;
 import uk.gov.justice.laa.dstew.access.model.CategoryOfLaw;
-import uk.gov.justice.laa.dstew.access.model.DecisionStatus;
 import uk.gov.justice.laa.dstew.access.model.IndividualType;
 import uk.gov.justice.laa.dstew.access.model.MatterType;
-import uk.gov.justice.laa.dstew.access.model.MeritsDecisionStatus;
 
 /** Converts between domain records and JPA entities for the createApplication use case. */
 public class ApplicationGatewayMapper {
@@ -147,53 +141,6 @@ public class ApplicationGatewayMapper {
         .justification(meritsDecision.getJustification())
         .modifiedAt(meritsDecision.getModifiedAt())
         .build();
-  }
-
-  /**
-   * Mutates the managed {@link ApplicationEntity} in-place from the supplied domain. Does NOT
-   * create a new entity — this preserves the JPA-managed {@literal @}Version and all audit fields
-   * on the application.
-   *
-   * <p>Updates: {@code decision} (build-or-update), {@code isAutoGranted}, {@code modifiedAt}, and
-   * each proceeding's {@code meritsDecision}. The domain is expected to contain the complete merged
-   * set of proceedings (with merits decisions already applied by the use case).
-   *
-   * @param entity the managed JPA entity to mutate
-   * @param domain the updated domain holding the new decision and fully merged proceedings
-   */
-  public void applyDecisionToEntity(ApplicationEntity entity, ApplicationDomain domain) {
-    DecisionEntity decisionEntity =
-        Optional.ofNullable(entity.getDecision()).orElseGet(() -> DecisionEntity.builder().build());
-    decisionEntity.setOverallDecision(DecisionStatus.valueOf(domain.decision().overallDecision()));
-    decisionEntity.setModifiedAt(domain.decision().modifiedAt());
-    entity.setDecision(decisionEntity);
-
-    entity.setIsAutoGranted(domain.isAutoGranted());
-    entity.setModifiedAt(Instant.now());
-
-    if (entity.getProceedings() != null && domain.proceedings() != null) {
-      Map<UUID, ProceedingEntity> proceedingEntityMap =
-          entity.getProceedings().stream()
-              .collect(Collectors.toMap(ProceedingEntity::getId, p -> p));
-      domain
-          .proceedings()
-          .forEach(
-              proceedingDomain -> {
-                ProceedingEntity proceedingEntity = proceedingEntityMap.get(proceedingDomain.id());
-                if (proceedingEntity != null && proceedingDomain.meritsDecision() != null) {
-                  MeritsDecisionEntity meritsDecision =
-                      Optional.ofNullable(proceedingEntity.getMeritsDecision())
-                          .orElseGet(MeritsDecisionEntity::new);
-                  meritsDecision.setDecision(
-                      MeritsDecisionStatus.valueOf(proceedingDomain.meritsDecision().decision()));
-                  meritsDecision.setReason(proceedingDomain.meritsDecision().reason());
-                  meritsDecision.setJustification(
-                      proceedingDomain.meritsDecision().justification());
-                  meritsDecision.setModifiedAt(proceedingDomain.meritsDecision().modifiedAt());
-                  proceedingEntity.setMeritsDecision(meritsDecision);
-                }
-              });
-    }
   }
 
   // ── Private helpers ──────────────────────────────────────────────────────
