+++ data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/usecase/shared/infrastructure/ApplicationGateway.java
@@ -42,12 +42,4 @@ public interface ApplicationGateway {
    * @return a list of IDs that could not be found (empty if all were present)
    */
   List<UUID> findMissingApplyApplicationIds(List<UUID> applyApplicationIds);
-
-  /**
-   * Persists decision and merits-decision changes to an existing managed application (UPDATE path).
-   * Implementations must reload the managed entity to avoid {@literal @}Version null.
-   *
-   * @param domain the updated application domain (id must be non-null)
-   */
-  void updateDecision(ApplicationDomain domain);
 }
